+++ spec/Config/ConfigManagerSpec.php
@@ -5,7 +5,9 @@ declare(strict_types = 1);
 namespace spec\hanneskod\readmetester\Config;
 
 use hanneskod\readmetester\Config\ConfigManager;
+use hanneskod\readmetester\Config\Configs;
 use hanneskod\readmetester\Config\RepositoryInterface;
+use hanneskod\readmetester\Config\Suite;
 use PhpSpec\ObjectBehavior;
 
 class ConfigManagerSpec extends ObjectBehavior
@@ -22,81 +24,426 @@ class ConfigManagerSpec extends ObjectBehavior
         $this->shouldHaveType(ConfigManager::class);
     }
 
-    function it_collects_repository_names(RepositoryInterface $repo)
+    function it_collects_repository_names($repo)
     {
         $repo->getConfigs()->willReturn([]);
         $repo->getRepositoryName()->willReturn('foo');
         $this->getLoadedRepositoryNames()->shouldIterateAs(['foo']);
     }
 
-    function it_ignores_empty_names(RepositoryInterface $repo)
+    function it_ignores_empty_names($repo)
     {
         $repo->getConfigs()->willReturn([]);
         $repo->getRepositoryName()->willReturn('');
         $this->getLoadedRepositoryNames()->shouldIterateAs([]);
     }
 
-    function it_can_read_config(RepositoryInterface $repo)
+    function it_finds_bootstrap($repo)
     {
-        $repo->getConfigs()->willReturn(['foo' => 'bar']);
-        $this->getConfig('foo')->shouldReturn('bar');
+        $repo->getConfigs()->willReturn([
+            Configs::BOOTSTRAP => __FILE__,
+        ]);
+
+        $this->getBootstrap()->shouldReturn(__FILE__);
     }
 
-    function it_can_read_nested_config(RepositoryInterface $repo)
+    function it_reas_bootstrap_from_last_loaded_repo($repo, RepositoryInterface $repoB)
     {
-        $repo->getConfigs()->willReturn(['foo' => ['bar' => 'foobar']]);
-        $this->getConfig('foo', 'bar')->shouldReturn('foobar');
-    }
+        $repo->getConfigs()->willReturn([
+            Configs::BOOTSTRAP => 'does-not-exist',
+        ]);
 
-    function it_load_multiple_repos($repo, RepositoryInterface $repoB)
-    {
-        $repo->getConfigs()->willReturn(['foo' => 'A', 'bar' => 'A']);
-        $repoB->getConfigs()->willReturn(['foo' => 'B']);
-        $repoB->getRepositoryName()->willReturn('');
+        $repoB->getConfigs()->willReturn([
+            Configs::BOOTSTRAP => __FILE__,
+        ]);
+
         $this->loadRepository($repoB);
-        $this->getConfig('foo')->shouldReturn('B');
-        $this->getConfig('bar')->shouldReturn('A');
+
+        $this->getBootstrap()->shouldReturn(__FILE__);
     }
 
-    function it_throws_on_missing_config(RepositoryInterface $repo)
+    function it_throws_on_unknown_bootstrap($repo)
+    {
+        $repo->getConfigs()->willReturn([
+            Configs::BOOTSTRAP => 'this-file-does-not-exist',
+        ]);
+
+        $this->shouldThrow(\RuntimeException::class)->duringGetBootstrap();
+    }
+
+    function it_throws_on_invalid_bootstrap($repo)
+    {
+        $repo->getConfigs()->willReturn([
+            Configs::BOOTSTRAP => ['this-is-not-scalar'],
+        ]);
+
+        $this->shouldThrow(\RuntimeException::class)->duringGetBootstrap();
+    }
+
+    function it_does_not_throw_on_empty_bootstrap($repo)
     {
         $repo->getConfigs()->willReturn([]);
-        $this->shouldThrow(\RuntimeException::class)->duringGetConfig('does-not-exist');
+
+        $this->getBootstrap()->shouldReturn('');
     }
 
-    function it_throws_on_non_scalar_value(RepositoryInterface $repo)
-    {
-        $repo->getConfigs()->willReturn(['not-scalar' => []]);
-        $this->shouldThrow(\RuntimeException::class)->duringGetConfig('not-scalar');
-    }
-
-    function it_casts_scalar_values_to_string(RepositoryInterface $repo)
-    {
-        $repo->getConfigs()->willReturn(['no-string' => 123]);
-        $this->getConfig('no-string')->shouldReturn('123');
-    }
-
-    function it_can_read_config_list(RepositoryInterface $repo)
-    {
-        $repo->getConfigs()->willReturn(['foo' => ['bar', 'baz']]);
-        $this->getConfigList('foo')->shouldIterateAs(['bar', 'baz']);
-    }
-
-    function it_throws_on_missing_config_list(RepositoryInterface $repo)
+    function it_reas_bootstrap_after_rebuild($repo, RepositoryInterface $repoB)
     {
         $repo->getConfigs()->willReturn([]);
-        $this->shouldThrow(\RuntimeException::class)->duringGetConfigList('does-not-exist');
+
+        $this->getBootstrap()->shouldReturn('');
+
+        $repoB->getConfigs()->willReturn([
+            Configs::BOOTSTRAP => __FILE__,
+        ]);
+
+        $this->loadRepository($repoB);
+
+        $this->getBootstrap()->shouldReturn(__FILE__);
     }
 
-    function it_throws_on_non_scalar_config_list_item(RepositoryInterface $repo)
+    function it_reads_cli_bootstrap_over_standard($repo)
     {
-        $repo->getConfigs()->willReturn(['foo' => [[]]]);
-        $this->shouldThrow(\RuntimeException::class)->duringGetConfigList('foo');
+        $repo->getConfigs()->willReturn([
+            Configs::BOOTSTRAP => 'this-file-does-not-exist',
+            Configs::CLI => [
+                Configs::BOOTSTRAP => __FILE__,
+            ]
+        ]);
+
+        $this->getBootstrap()->shouldReturn(__FILE__);
     }
 
-    function it_casts_scalar_config_list_items_to_string(RepositoryInterface $repo)
+    function it_finds_subscribers($repo)
     {
-        $repo->getConfigs()->willReturn(['foo' => [1234]]);
-        $this->getConfigList('foo')->shouldIterateAs(['1234']);
+        $repo->getConfigs()->willReturn([
+            Configs::SUBSCRIBERS => ['foo'],
+            Configs::OUTPUT => 'bar'
+        ]);
+
+        $this->getSubscribers()->shouldYieldValues(['foo', 'bar']);
+    }
+
+    function it_expands_output_name($repo)
+    {
+        $repo->getConfigs()->willReturn([
+            Configs::SUBSCRIBERS => [],
+            Configs::OUTPUT => 'debug'
+        ]);
+
+        $this->getSubscribers()->shouldYieldValues([Configs::OUTPUT_ID[Configs::OUTPUT_ID_DEBUG]]);
+    }
+
+    function it_reads_cli_subscribers_over_standard($repo)
+    {
+        $repo->getConfigs()->willReturn([
+            Configs::CLI => [
+                Configs::OUTPUT => 'bar',
+            ]
+        ]);
+
+        $this->getSubscribers()->shouldYieldValues(['bar']);
+    }
+
+    function it_throws_on_unknown_suite($repo)
+    {
+        $this->shouldThrow(\RuntimeException::class)->duringGetSuite('does-not-exist');
+    }
+
+    function it_returns_named_suite($repo)
+    {
+        $repo->getConfigs()->willReturn([
+            Configs::SUITES => [
+                'foosuite' => []
+            ]
+        ]);
+
+        $this->getSuite('foosuite')->shouldBeLike(
+            new Suite(name: 'foosuite')
+        );
+    }
+
+    function it_returns_named_suite_even_if_not_active($repo)
+    {
+        $repo->getConfigs()->willReturn([
+            Configs::SUITES => [
+                'not-active' => [
+                    Configs::ACTIVE => false,
+                ]
+            ]
+        ]);
+
+        $this->getSuite('not-active')->shouldBeLike(
+            new Suite(
+                name: 'not-active',
+                active: false
+            )
+        );
+    }
+
+    function it_adds_defaults_to_suite($repo)
+    {
+        $repo->getConfigs()->willReturn([
+            Configs::DEFAULTS => [
+                    Configs::INCLUDE_PATHS => ['default'],
+            ],
+            Configs::SUITES => [
+                'foosuite' => [],
+            ],
+        ]);
+
+        $this->getSuite('foosuite')->shouldBeLike(
+            new Suite(
+                name: 'foosuite',
+                includePaths: ['default'],
+            )
+        );
+    }
+
+    function it_overwrites_defaults_with_suite_data($repo)
+    {
+        $repo->getConfigs()->willReturn([
+            Configs::DEFAULTS => [
+                    Configs::FILE_EXTENSIONS => ['default'],
+            ],
+            Configs::SUITES => [
+                'foosuite' => [
+                    Configs::FILE_EXTENSIONS => ['suite'],
+                ],
+            ],
+        ]);
+
+        $this->getSuite('foosuite')->shouldBeLike(
+            new Suite(
+                name: 'foosuite',
+                fileExtensions: ['suite'],
+            )
+        );
+    }
+
+    function it_overwrites_suite_data_with_cli_data($repo)
+    {
+        $repo->getConfigs()->willReturn([
+            Configs::DEFAULTS => [
+                    Configs::EXCLUDE_PATHS => ['default'],
+            ],
+            Configs::SUITES => [
+                'foosuite' => [
+                    Configs::EXCLUDE_PATHS => ['suite'],
+                ],
+            ],
+            Configs::CLI => [
+                    Configs::EXCLUDE_PATHS => ['cli'],
+            ],
+        ]);
+
+        $this->getSuite('foosuite')->shouldBeLike(
+            new Suite(
+                name: 'foosuite',
+                excludePaths: ['cli'],
+            )
+        );
+    }
+
+    function it_applies_defaults_and_cli_data_to_all_suites($repo)
+    {
+        $repo->getConfigs()->willReturn([
+            Configs::CLI => [
+                    Configs::INCLUDE_PATHS => ['cli_include'],
+            ],
+            Configs::SUITES => [
+                'foosuite' => [
+                    Configs::FILE_EXTENSIONS => ['foosuite_extension'],
+                ],
+                'barsuite' => [
+                    Configs::FILE_EXTENSIONS => ['barsuite_extension'],
+                ],
+            ],
+            Configs::DEFAULTS => [
+                    Configs::EXCLUDE_PATHS => ['default_exclude'],
+            ],
+        ]);
+
+        $this->getAllSuites()->shouldIterateLike([
+            new Suite(
+                name: 'foosuite',
+                excludePaths: ['default_exclude'],
+                includePaths: ['cli_include'],
+                fileExtensions: ['foosuite_extension'],
+            ),
+            new Suite(
+                name: 'barsuite',
+                excludePaths: ['default_exclude'],
+                includePaths: ['cli_include'],
+                fileExtensions: ['barsuite_extension'],
+            ),
+        ]);
+    }
+
+    function it_ignores_non_active_suites_when_accessing_all_suites($repo)
+    {
+        $repo->getConfigs()->willReturn([
+            Configs::SUITES => [
+                'active' => [
+                    Configs::ACTIVE => true,
+                ],
+                'not-active' => [
+                    Configs::ACTIVE => false,
+                ],
+            ],
+        ]);
+
+        $this->getAllSuites()->shouldYieldValues([
+            new Suite(name: 'active')
+        ]);
+    }
+
+    function it_creates_default_suite_if_non_is_definied($repo)
+    {
+        $repo->getConfigs()->willReturn([]);
+
+        $this->getAllSuites()->shouldIterateLike([
+            new Suite(name: 'default')
+        ]);
+    }
+
+    function it_loads_complex_data_from_multiple_repos(
+        RepositoryInterface $defaults,
+        RepositoryInterface $cli,
+        RepositoryInterface $user,
+    ) {
+        $defaults->getConfigs()->willReturn([
+            Configs::DEFAULTS => [
+                    Configs::EXCLUDE_PATHS => ['default_exclude'],
+                    Configs::FILE_EXTENSIONS => ['default_extension'],
+            ],
+        ]);
+
+        $cli->getConfigs()->willReturn([
+            Configs::CLI => [
+                    Configs::INCLUDE_PATHS => ['cli_include'],
+            ],
+        ]);
+
+        $user->getConfigs()->willReturn([
+            Configs::SUITES => [
+                'foosuite' => [
+                    // Should overwrite default_extension
+                    Configs::FILE_EXTENSIONS => ['foosuite_extension'],
+                ],
+                'barsuite' => [
+                    // Should be overwritten by cli_include
+                    Configs::INCLUDE_PATHS => ['bars_include'],
+                ],
+            ],
+        ]);
+
+        $this->loadRepository($cli);
+        $this->loadRepository($user);
+        $this->loadRepository($defaults);
+
+        $this->getAllSuites()->shouldIterateLike([
+            new Suite(
+                name: 'foosuite',
+                excludePaths: ['default_exclude'],
+                includePaths: ['cli_include'],
+                fileExtensions: ['foosuite_extension'],
+            ),
+            new Suite(
+                name: 'barsuite',
+                excludePaths: ['default_exclude'],
+                includePaths: ['cli_include'],
+                fileExtensions: ['default_extension'],
+            ),
+        ]);
+    }
+
+    function it_readbuilds_suites_after_repo_load($repo, RepositoryInterface $foosuite)
+    {
+        $repo->getConfigs()->willReturn([]);
+
+        $foosuite->getConfigs()->willReturn([
+            Configs::SUITES => [
+                'foosuite' => []
+            ],
+        ]);
+
+        $this->getAllSuites()->shouldIterateLike([
+            new Suite(name: 'default')
+        ]);
+
+        $this->loadRepository($foosuite);
+
+        $this->getAllSuites()->shouldIterateLike([
+            new Suite(name: 'foosuite')
+        ]);
+    }
+
+    function it_combines_multiple_defaults_definitions(
+        RepositoryInterface $defaultsA,
+        RepositoryInterface $defaultsB,
+        RepositoryInterface $foosuite,
+    ) {
+        $defaultsA->getConfigs()->willReturn([
+            Configs::DEFAULTS => [
+                // Should be overwritten by defaultsB
+                Configs::EXCLUDE_PATHS => ['defaultsA_exclude'],
+                // Should be overwritten by foosuite_defaults
+                Configs::INCLUDE_PATHS => ['defaultsA_include'],
+                // Should be overwritten by foosuite
+                Configs::INPUT_LANGUAGE => 'defaultsA_input_language',
+                // SHould stand
+                Configs::RUNNER => 'defaultsA_runner',
+            ],
+        ]);
+
+        $defaultsB->getConfigs()->willReturn([
+            Configs::DEFAULTS => [
+                Configs::EXCLUDE_PATHS => ['defaultsB_exclude'],
+            ],
+        ]);
+
+        $foosuite->getConfigs()->willReturn([
+            Configs::DEFAULTS => [
+                Configs::INCLUDE_PATHS => ['foosuite_defaults_include'],
+            ],
+            Configs::SUITES => [
+                'foosuite' => [
+                    Configs::INPUT_LANGUAGE => 'foosuite_input_language',
+                ]
+            ],
+        ]);
+
+        $this->loadRepository($defaultsA);
+        $this->loadRepository($defaultsB);
+        $this->loadRepository($foosuite);
+
+        $this->getAllSuites()->shouldIterateLike([
+            new Suite(
+                name: 'foosuite',
+                excludePaths: ['defaultsB_exclude'],
+                includePaths: ['foosuite_defaults_include'],
+                inputLanguage: 'foosuite_input_language',
+                runner: 'defaultsA_runner',
+            )
+        ]);
+    }
+
+    function getMatchers(): array
+    {
+        return [
+            'yieldValues' => function (\Generator $generator, array $expected) {
+                $generatorValues = array_values(iterator_to_array($generator, false));
+                $expectedValues = array_values($expected);
+
+                foreach ($generatorValues as $key => $value) {
+                    if (!isset($expectedValues[$key]) || $value != $expectedValues[$key]) {
+                        return false;
+                    }
+                }
+
+                return true;
+            }
+        ];
     }
 }