+++ src/Runner/RunnerInterface.php
@@ -3,7 +3,6 @@
 namespace hanneskod\readmetester\Runner;
 
 use hanneskod\readmetester\Example\ExampleObj;
-use hanneskod\readmetester\Utils\CodeBlock;
 
 /**
  * Defines object that are able to execute example code
@@ -12,5 +11,5 @@ interface RunnerInterface
 {
     public function run(ExampleObj $example): OutcomeInterface;
 
-    public function setBootstrap(CodeBlock $bootstrap): void;
+    public function setBootstrap(string $filename): void;
 }