+++ CHANGELOG.md
@@ -10,13 +10,13 @@ This project now adheres to [Semantic Versioning](http://semver.org/).
 - Removed the `--ignore-unknown-annotations` command line option
 - Removed the `--filter` command line option
 - Added the `--file-extension` command line option
-- Added the `--ignore` command line option
+- Added the `--exclude` command line option
 - Added the `--stop-on-failure` command line option
 - Added the `--input` command line option
 - Added the `--config` command line option
 - Renamed command line option `--format` > `--output`
 - Using php8 attributes instead of annotations
-- Added support for `readme-tester.yaml` config file
+- Added support for config files
 - Use a propper event dispatcher
 - Added multiple new attributes, se [docs/attribute_reference.md](docs/attribute_reference.md)